move libSVMFile to MLUtils and rename to loadLibSVMData
