--- mllib/src/test/scala/org/apache/spark/mllib/MLContextSuite.scala
@@ -1,64 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.mllib
-
-import java.io.File
-
-import org.scalatest.FunSuite
-
-import com.google.common.base.Charsets
-import com.google.common.io.Files
-
-import org.apache.spark.mllib.linalg.Vectors
-import org.apache.spark.mllib.util.LocalSparkContext
-
-class MLContextSuite extends FunSuite with LocalSparkContext {
-  test("libSVMFile") {
-    val lines =
-      """
-        |1 1:1.0 3:2.0 5:3.0
-        |0
-        |0 2:4.0 4:5.0 6:6.0
-      """.stripMargin
-    val tempDir = Files.createTempDir()
-    val file = new File(tempDir.getPath, "part-00000")
-    Files.write(lines, file, Charsets.US_ASCII)
-
-    val mlc = MLContext(sc)
-
-    val pointsWithNumFeatures = mlc.libSVMFile(tempDir.toURI.toString, numFeatures = 6).collect()
-    val pointsWithoutNumFeatures = mlc.libSVMFile(tempDir.toURI.toString).collect()
-
-    for (points <- Seq(pointsWithNumFeatures, pointsWithoutNumFeatures)) {
-      assert(points.length === 3)
-      assert(points(0).label === 1.0)
-      assert(points(0).features === Vectors.sparse(6, Seq((0, 1.0), (2, 2.0), (4, 3.0))))
-      assert(points(1).label == 0.0)
-      assert(points(1).features == Vectors.sparse(6, Seq()))
-      assert(points(2).label === 0.0)
-      assert(points(2).features === Vectors.sparse(6, Seq((1, 4.0), (3, 5.0), (5, 6.0))))
-    }
-
-    try {
-      file.delete()
-      tempDir.delete()
-    } catch {
-      case t: Throwable =>
-    }
-  }
-}
